+++ Eval1/src/main/java/com/eval1/models/rate/Rate.java
@@ -21,6 +21,6 @@ import java.lang.Integer;
 public class Rate extends HasId {
 
 	private Date date;
-	private BigDecimal rate;
+	private Double rate;
 
 }
